+++ search/tabs.js
@@ -1,69 +1,35 @@
 const resultsTabsList = document.getElementById("tabs-list");
+
 let tabsObjects = [];
 let activeIndex = 0;
+let scoreThreshold = 0;
 
-window.browser = (function () {
-  return window.msBrowser || window.browser || window.chrome;
-})();
-
-function setOptions(result) {
-  bookmarkBackgroundColor = result.bookmarkBackgroundColor;
-  bookmarkTextColor = result.bookmarkTextColor;
-  // bookmarkBorderColor = result.bookmarkBorderColor;
-
-  openTabsBackgroundColor = result.openTabsBackgroundColor;
-  openTabsTextColor = result.openTabsTextColor;
-  // openTabsBorderColor = result.openTabsBorderColor;
-
-  focusItemBackgroundColor = result.focusItemBackgroundColor;
-  focusItemTextColor = result.focusItemTextColor;
-  focusItemBorderColor = result.focusItemBorderColor;
-
+browser.storage.sync.get("scoreThreshold").then((result) => {
   scoreThreshold = result.scoreThreshold;
-}
+});
 
-const config = browser.storage.sync.get([
-  "bookmarkBackgroundColor",
-  "bookmarkTextColor",
-  // "bookmarkBorderColor",
-
-  "openTabsBackgroundColor",
-  "openTabsTextColor",
-  // "openTabsBorderColor",
-
-  "focusItemBackgroundColor",
-  "focusItemTextColor",
-  "focusItemBorderColor",
-  "scoreThreshold",
-]);
-
-config.then(setOptions);
-
-function searchResultTabs() {
+function searchResultTabs(e) {
   activeIndex = 0;
   resultsTabsList.textContent = "";
-  let searchQuery = document.getElementById("find-input").value;
 
-  // Get matching tabs TODO Add option to customize threshold for scores
-  console.log(scoreThreshold);
+  const searchQuery = e.target.value;
+
+  // Get matching tabs
   fuzzyResultsAll = fuzzysort.go(
     searchQuery,
     tabsObjects.map((a) => a.title),
     { threshold: scoreThreshold },
   );
 
-  for (i in fuzzyResultsAll) {
-    currentTab =
-      tabsObjects[
-        tabsObjects.findIndex((x) => x.title === fuzzyResultsAll[i].target)
-      ];
+  for (result of fuzzyResultsAll) {
+    const currentTab =
+      tabsObjects[tabsObjects.findIndex((x) => x.title === result.target)];
 
     if (currentTab == null) {
       continue;
     }
 
-    let link = new TabLink(currentTab);
-    tabLink = link.getLink();
+    const tabLink = createListItem(currentTab);
 
     if (resultsTabsList.childElementCount === 0) {
       tabLink.classList.add("active");
@@ -71,114 +37,62 @@ function searchResultTabs() {
 
     resultsTabsList.appendChild(tabLink);
   }
-
-  // Switch to first item with Enter (ie. don't put <CR> into text box)
-  document.getElementById("find-input").onkeydown = function (e) {
-    if (e.code === "ArrowUp") {
-      e.preventDefault();
-      if (tabsObjects.length === 0) {
-        return;
-      }
-
-      upateSelection(activeIndex > 0 ? activeIndex - 1 : 0);
-    }
-
-    if (e.code === "ArrowDown") {
-      e.preventDefault();
-      if (tabsObjects.length === 0) {
-        return;
-      }
-
-      const maxIndex = resultsTabsList.childElementCount - 1;
-      upateSelection(activeIndex < maxIndex ? activeIndex + 1 : maxIndex);
-    }
-
-    if (e.code === "Enter") {
-      e.preventDefault();
-      resultsTabsList.children[activeIndex].click();
-      // Close popup after switching tabs
-      window.close();
-    }
-  };
 }
 
-document.addEventListener("DOMContentLoaded", function (e) {
-  getTabs();
-  getBookmarks();
-
-  // Do this whenever the text box changes
+document.addEventListener("DOMContentLoaded", function (_) {
   const input = document.getElementById("find-input");
-  if (input != null) {
-    input.oninput = searchResultTabs;
-    // Focus on the input box on load
-    document.getElementById("find-input").focus();
-  }
 
-  // Click on link to switch to that tab/window
-  document.addEventListener("click", (e) => {
-    const link =
-      e.target.tagName.toLowerCase() == "li"
-        ? e.target.firstElementChild
-        : e.target;
+  input.oninput = searchResultTabs;
+  input.onkeydown = handleKeymaps;
 
-    const openTabId = +link.getAttribute("href");
-    const winId = +link.getAttribute("window");
-    const bookmarkUrl = link.href;
+  // NOTE: Focusing is unreliable, trying multiple times.
+  focusInput();
+  setTimeout(focusInput, 50);
+  setTimeout(focusInput, 100);
 
-    if (Number.isInteger(openTabId)) {
-      browser.windows.update(winId, {
-        focused: true,
-      });
-      browser.tabs.update(openTabId, {
-        active: true,
-      });
-    } else {
-      browser.tabs.create({ url: bookmarkUrl });
-    }
-    e.preventDefault();
-  });
+  getAllTabs().then((tabs) => (tabsObjects = tabs));
+  // getAllBookmarks().then((tabs) => (tabsObjects = tabs));
 });
 
-function getBookmarks() {
-  function logItems(bookmarkItem) {
-    if (bookmarkItem.url) {
-      tabsObjects.push({
-        title: bookmarkItem.title,
-        url: bookmarkItem.url,
-        bookmark: 1,
-      });
-    }
-    if (bookmarkItem.children) {
-      for (child of bookmarkItem.children) {
-        logItems(child);
+const focusInput = () => {
+  const input = document.getElementById("find-input");
+
+  input.blur();
+  input.focus();
+  input.click();
+};
+
+function handleKeymaps(e) {
+  switch (e.code) {
+    case "ArrowUp":
+      e.preventDefault();
+      if (resultsTabsList.childElementCount === 0) {
+        return;
       }
-    }
-  }
+      setActiveIndex(activeIndex > 0 ? activeIndex - 1 : 0);
+      break;
 
-  function logTree(bookmarkItems) {
-    logItems(bookmarkItems[0]);
-  }
+    case "ArrowDown":
+      e.preventDefault();
+      if (resultsTabsList.childElementCount === 0) {
+        return;
+      }
+      const maxIndex = resultsTabsList.childElementCount - 1;
+      setActiveIndex(activeIndex < maxIndex ? activeIndex + 1 : maxIndex);
+      break;
 
-  browser.bookmarks.getTree(logTree);
+    case "Enter":
+      e.preventDefault();
+      if (resultsTabsList.childElementCount === 0) {
+        return;
+      }
+      resultsTabsList.children[activeIndex].click();
+      window.close();
+      break;
+  }
 }
 
-function getTabs() {
-  tabsObjects = [];
-  browser.tabs.query({}, function (tabs) {
-    for (let tab of tabs) {
-      // console.log(tab.title + " : "  + tab.id + " : " + tab.windowId);
-      tabsObjects.push({
-        title: tab.title,
-        url: tab.url,
-        id: tab.id,
-        winId: tab.windowId,
-        bookmark: 0,
-      });
-    }
-  });
-}
-
-function upateSelection(newIndex) {
+function setActiveIndex(newIndex) {
   if (activeIndex != newIndex) {
     resultsTabsList.children[activeIndex].classList.remove("active");
   }
@@ -186,46 +100,82 @@ function upateSelection(newIndex) {
   resultsTabsList.children[activeIndex].classList.add("active");
 }
 
-function setActive(event) {
-  event.target.classList.add("active");
+const setLinkActive = (e) => e.target.classList.add("active");
+const setLinkInactive = (e) => e.target.classList.remove("active");
+
+const openLink = (e) => {
+  e.preventDefault();
+
+  const link =
+    e.target.nodeName.toLowerCase() === "li"
+      ? e.target.firstElementChild
+      : e.target;
+
+  const href = link.getAttribute("href");
+  const winId = parseInt(link.getAttribute("window"));
+  const tabId = parseInt(href);
+
+  if (Number.isInteger(tabId)) {
+    browser.windows.update(winId, { focused: true });
+    browser.tabs.update(tabId, { active: true });
+  } else {
+    browser.tabs.create({ url: href });
+  }
+};
+
+function createListItem(tab) {
+  const item = document.createElement("li");
+  item.classList.add("list-group-item");
+  item.onmouseenter = setLinkActive;
+  item.onfocus = setLinkActive;
+  item.onmouseleave = setLinkInactive;
+  item.onblur = setLinkInactive;
+  item.onclick = openLink;
+
+  const link = document.createElement("a");
+  link.classList.add("open-tab-link");
+  link.textContent = tab.title.slice(0, 64);
+  link.setAttribute("href", tab.tabId ?? tab.url);
+  link.setAttribute("window", tab.winId);
+
+  item.appendChild(link);
+
+  return item;
 }
 
-function setInactive(event) {
-  event.target.classList.remove("active");
+async function getAllTabs() {
+  return browser.tabs.query({}).then((tabs) =>
+    tabs.map((tab) => {
+      return {
+        title: tab.title,
+        url: tab.url,
+        tabId: tab.id,
+        winId: tab.windowId,
+      };
+    }),
+  );
 }
 
-class TabLink {
-  constructor(tabObject) {
-    this.item = document.createElement("li");
-    this.item.classList.add("list-group-item");
-    this.item.onmouseenter = setActive;
-    this.item.onfocus = setActive;
-    this.item.onmouseleave = setInactive;
-    this.item.onblur = setInactive;
+async function getAllBookmarks() {
+  const uniqueUrls = new Set([]);
+  const bookmarks = [];
 
-    this.link = document.createElement("a");
-
-    if (tabObject.bookmark) {
-      this.link.classList.add("bookmark-link");
-      this.link.textContent = tabObject.title.slice(0, 64);
-      this.link.setAttribute("href", tabObject.url);
-      this.link.setAttribute("window", tabObject.winId);
-
-      this.link.style.backgroundColor = bookmarkBackgroundColor;
-      this.link.style.color = bookmarkTextColor;
-    } else if (!tabObject.bookmark) {
-      this.link.classList.add("open-tab-link");
-      this.link.textContent = tabObject.title.slice(0, 64);
-      this.link.setAttribute("href", tabObject.id);
-      this.link.setAttribute("window", tabObject.winId);
-
-      this.link.style.backgroundColor = openTabsBackgroundColor;
-      this.link.style.color = openTabsTextColor;
+  function walk(node) {
+    if (node.url != null && !uniqueUrls.has(node.url)) {
+      bookmarks.push({
+        title: node.title,
+        url: node.url,
+      });
     }
 
-    this.item.appendChild(this.link);
+    if (node.children != null && node.children.length > 0) {
+      for (child of node.children) {
+        walk(child);
+      }
+    }
   }
-  getLink() {
-    return this.item;
-  }
+
+  await browser.bookmarks.getTree().then((tree) => walk(tree[0]));
+
+  return bookmarks;
 }